refactor webfinger and pdf viewer app to defineWebApplication
