+++ packages/web-app-pdf-viewer/src/index.ts
@@ -1,41 +1,48 @@
+import { useGettext } from 'vue3-gettext'
 import translations from '../l10n/translations.json'
-import { AppWrapperRoute, $gettext } from '@opencloud-eu/web-pkg'
+import { AppWrapperRoute, defineWebApplication } from '@opencloud-eu/web-pkg'
 import PdfViewer from './App.vue'
 
-const routes = [
-  {
-    path: '/:driveAliasAndItem(.*)?',
-    component: AppWrapperRoute(PdfViewer, {
-      applicationId: 'pdf-viewer',
-      urlForResourceOptions: {
-        disposition: 'inline'
+export default defineWebApplication({
+  setup() {
+    const { $gettext } = useGettext()
+
+    const appId = 'pdf-viewer'
+
+    const routes = [
+      {
+        path: '/:driveAliasAndItem(.*)?',
+        component: AppWrapperRoute(PdfViewer, {
+          applicationId: appId,
+          urlForResourceOptions: {
+            disposition: 'inline'
+          }
+        }),
+        name: 'pdf-viewer',
+        meta: {
+          authContext: 'hybrid',
+          title: $gettext('PDF Viewer'),
+          patchCleanPath: true
+        }
       }
-    }),
-    name: 'pdf-viewer',
-    meta: {
-      authContext: 'hybrid',
-      title: $gettext('PDF Viewer'),
-      patchCleanPath: true
+    ]
+
+    return {
+      appInfo: {
+        name: $gettext('PDF Viewer'),
+        id: appId,
+        icon: 'resource-type-pdf',
+        iconFillType: 'fill',
+        iconColor: 'var(--oc-color-icon-pdf)',
+        extensions: [
+          {
+            extension: 'pdf',
+            routeName: 'pdf-viewer'
+          }
+        ]
+      },
+      routes,
+      translations
     }
   }
-]
-
-const appInfo = {
-  name: $gettext('PDF Viewer'),
-  id: 'pdf-viewer',
-  icon: 'resource-type-pdf',
-  iconFillType: 'fill',
-  iconColor: 'var(--oc-color-icon-pdf)',
-  extensions: [
-    {
-      extension: 'pdf',
-      routeName: 'pdf-viewer'
-    }
-  ]
-}
-
-export default {
-  appInfo,
-  routes,
-  translations
-}
+})
