+++ packages/web-pkg/src/router/index.ts
@@ -34,7 +34,8 @@ import {
   isLocationTrashActive,
   createLocationTrash
 } from './trash'
-import { isLocationActiveDirector, createLocation, $gettext } from './utils'
+import { isLocationActiveDirector, createLocation } from './utils'
+import { $gettext } from '../utils/dummyGettext'
 import type { ActiveRouteDirectorFunc } from './utils'
 
 const ROOT_ROUTE = {
