+++ packages/web-pkg/src/utils/dummyGettext.ts
@@ -1,0 +1,8 @@
+/**
+ * just a dummy function to trick gettext tools
+ *
+ * @param msg
+ */
+export function $gettext(msg: string): string {
+  return msg
+}
