+++ packages/web-pkg/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './encodePath'
 export * from './objectKeys'
 export * from './semver'
 export * from './types'
+export * from './dummyGettext'
